+++ web-service/.gitignore
@@ -1,2 +1,2 @@
-.pyc
-.DS_Store
+**/*.pyc
+**/.DS_Store
