+++ web-service/models.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python
+#
+# Copyright 2014 Google Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from google.appengine.ext import ndb
+
+class BaseModel(ndb.Model):
+    added_on = ndb.DateTimeProperty(auto_now_add = True)
+    updated_on = ndb.DateTimeProperty(auto_now = True)
+
+class SiteInformation(BaseModel):
+    url = ndb.TextProperty()
+    favicon_url = ndb.TextProperty()
+    title = ndb.TextProperty()
+    description = ndb.TextProperty()
+    jsonlds = ndb.TextProperty()
